Plugin: idah-video separate IndexedDB and prepare middleware
