+++ plugins/idah-video/frontend/src/lib/plugin/layout/sidebar/annotation-sidebar.svelte
@@ -14,10 +14,16 @@
   import { entryRoot } from "$lib/plugin/video-annotation-activity/store/idb-store.svelte";
   import { currentMode } from "$lib/plugin/video-annotation-activity/store/store";
 
-  import type { IActivityContext, IConfigValue } from "$idah/context/activity-context";
-  import type { AnnotationGroup, AnnotationValue } from "$idah/context/annotation-context";
+  import type {
+    IActivityContext,
+    IConfigValue,
+  } from "$idah/context/activity-context";
+  import type {
+    AnnotationGroup,
+    AnnotationValue,
+  } from "$idah/context/annotation-context";
   import type { VideoAnnotationObject } from "$lib/plugin/video-annotation-activity/context/video-annotation-context";
-  import type { AnnotationsIndexedDB } from "$lib/plugin/video-annotation-activity/indexedDB";
+  import type { AnnotationsMiddleware } from "$lib/plugin/video-annotation-activity/indexedDB.svelte";
 
   // Props
   let {
@@ -37,10 +43,12 @@
     annotationValue: AnnotationValue;
     onEditValue: (annotationValue: AnnotationValue, mode: string) => void;
     onSelectAnnotation: (annotation?: VideoAnnotationObject) => void;
-    onSelectAnnotationGroup: (annotationGroup: AnnotationGroup<VideoAnnotationObject>) => void;
+    onSelectAnnotationGroup: (
+      annotationGroup: AnnotationGroup<VideoAnnotationObject>,
+    ) => void;
     onDeleteAnnotation: (annotation: VideoAnnotationObject) => void;
     context: IActivityContext;
-    db?: AnnotationsIndexedDB;
+    db?: AnnotationsMiddleware;
     class?: string | null;
   } = $props();
 
@@ -105,7 +113,12 @@
   }
 </script>
 
-<Sidebar variant="inset" collapsible="none" class={cn(className)} style="width: ${sidebarWidthRem}rem">
+<Sidebar
+  variant="inset"
+  collapsible="none"
+  class={cn(className)}
+  style="width: ${sidebarWidthRem}rem"
+>
   {#if !tools.has($currentMode)}
     <SidebarHeader>
       <InputField
