+++ plugins/idah-video/frontend/src/lib/plugin/video-annotation-activity/indexedDB.ts
@@ -1,0 +1,313 @@
+import {
+  type VideoAnnotationObject,
+  type VideoFrameSelection,
+} from "$lib/plugin/video-annotation-activity/context/video-annotation-context";
+
+// the current version of IndexedDB, bump incrementally if there's a change
+export const currentDBVersion: number = 2;
+
+export const storeDefinition: StoresDefinition = {
+  annotations: [
+    { index: "start", path: "shape.start", opts: { unique: false } },
+    { index: "end", path: "shape.end", opts: { unique: false } },
+    { index: "range", path: "shape.range", opts: { unique: false } },
+    { index: "type", path: "shape.type", opts: { unique: false } },
+    { index: "category", path: "value.category", opts: { unique: false } },
+    { index: "created_at", path: "metadata.updatedAt", opts: { unique: false } },
+    { index: "updated_at", path: "metadata.createdAt", opts: { unique: false } },
+    { index: "groupIdIndex", path: "metadata.metadata.group_id", opts: { unique: false } },
+  ],
+  keyframes: [
+    { index: "frame", path: "frame", opts: { unique: false } },
+    { index: "annotation", path: "annotation", opts: { unique: false } },
+  ],
+};
+
+export type StoreDefinition = {
+  index: string;
+  path: string;
+  opts?: IDBIndexParameters;
+};
+
+export type StoresDefinition = {
+  [name: string]: StoreDefinition[];
+};
+
+/**
+ * Pure IndexedDB database layer
+ * Handles only CRUD operations, no business logic or state management
+ */
+export class AnnotationsIndexedDB {
+  private db: IDBDatabase;
+
+  constructor(db: IDBDatabase) {
+    this.db = db;
+  }
+
+  /**
+   * Get a single item by key
+   */
+  async get(store_name: string, key: string): Promise<VideoAnnotationObject> {
+    return new Promise<VideoAnnotationObject>((resolve, reject) => {
+      const transaction = this.db.transaction(store_name, "readonly");
+      const store = transaction.objectStore(store_name);
+      const request = store.get(key);
+
+      request.onsuccess = (_) => resolve(request.result);
+      request.onerror = (_) => reject(request.error);
+    });
+  }
+
+  /**
+   * Get all items from a store
+   */
+  async getAll(storename: string): Promise<VideoAnnotationObject[]> {
+    return new Promise<VideoAnnotationObject[]>((resolve, reject) => {
+      const transaction = this.db.transaction(storename, "readonly");
+      const store = transaction.objectStore(storename);
+      const request = store.getAll();
+
+      request.onsuccess = () => resolve(request.result);
+      request.onerror = () => reject(request.error);
+    });
+  }
+
+  /**
+   * Get all items by index
+   */
+  async getAllByIndex(key: string, value?: string): Promise<VideoAnnotationObject[]> {
+    return new Promise<VideoAnnotationObject[]>((resolve, reject) => {
+      const transaction = this.db.transaction("annotations", "readonly");
+      const store = transaction.objectStore("annotations").index(key);
+      const keyRange = value != undefined ? IDBKeyRange.only(value) : undefined;
+      const request = store.getAll(keyRange);
+
+      request.onsuccess = () => resolve(request.result);
+      request.onerror = () => reject(request.error);
+    });
+  }
+
+  /**
+   * Put (upsert) an annotation
+   */
+  async putAnnotation(annotation: VideoAnnotationObject): Promise<void> {
+    const transaction = this.db.transaction(["annotations", "keyframes"], "readwrite");
+    const astore = transaction.objectStore("annotations");
+    const kstore = transaction.objectStore("keyframes");
+
+    astore.put(
+      {
+        ...annotation,
+        value: {
+          ...annotation.value,
+          category: annotation.value.category || "null",
+        },
+      },
+      annotation.metadata.id,
+    );
+
+    annotation.shape.frames?.forEach((k: VideoFrameSelection) => {
+      kstore.put({ annotation: annotation.metadata.id, ...k }, [annotation.metadata.id, k.frame]);
+    });
+
+    return new Promise<void>((resolve, reject) => {
+      transaction.oncomplete = () => resolve();
+      transaction.onerror = () => reject();
+    });
+  }
+
+  /**
+   * Put (upsert) multiple annotations
+   */
+  async putAnnotations(annotations: VideoAnnotationObject[]): Promise<void> {
+    const transaction = this.db.transaction(["annotations", "keyframes"], "readwrite");
+    const astore = transaction.objectStore("annotations");
+    const kstore = transaction.objectStore("keyframes");
+
+    annotations.forEach((annotation) => {
+      console.debug({ IDBtype: "put", annotation });
+      astore.put(
+        {
+          ...annotation,
+          value: {
+            ...annotation.value,
+            category: annotation.value.category || "null",
+          },
+        },
+        annotation.metadata.id,
+      );
+
+      annotation.shape.frames?.forEach((k: VideoFrameSelection) => {
+        kstore.put({ annotation: annotation.metadata.id, ...k }, [annotation.metadata.id, k.frame]);
+      });
+    });
+
+    return new Promise<void>((resolve, reject) => {
+      transaction.oncomplete = () => resolve();
+      transaction.onerror = () => reject();
+    });
+  }
+
+  /**
+   * Delete an annotation and its keyframes
+   */
+  async deleteAnnotation(annotation_id: string): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+      const ktransaction = this.db.transaction("keyframes", "readwrite");
+      const kstore = ktransaction.objectStore("keyframes");
+      const range = IDBKeyRange.bound([annotation_id, 0], [annotation_id + " ", 0]);
+      const request = kstore.delete(range);
+
+      request.onsuccess = (_) => {
+        const atransaction = this.db.transaction("annotations", "readwrite");
+        const astore = atransaction.objectStore("annotations");
+        const deleteRequest = astore.delete(annotation_id);
+
+        deleteRequest.onsuccess = () => resolve();
+        deleteRequest.onerror = () => reject(deleteRequest.error);
+      };
+      request.onerror = (_) => reject(request.error);
+    });
+  }
+
+  /**
+   * Add a keyframe to an annotation
+   */
+  async addKeyFrame(annotation_id: string, keyFrame: VideoFrameSelection): Promise<void> {
+    const transaction = this.db.transaction(["annotations", "keyframes"], "readwrite");
+    const kstore = transaction.objectStore("keyframes");
+    const astore = transaction.objectStore("annotations");
+
+    kstore.put({ annotation: annotation_id, ...keyFrame }, [annotation_id, keyFrame.frame]);
+
+    return new Promise<void>((resolve, reject) => {
+      transaction.oncomplete = async () => {
+        // Fetch and update the annotation with all keyframes
+        const annotation = (await this.get("annotations", annotation_id)) as VideoAnnotationObject;
+        const keyframes = await this.getKeyFrames(annotation_id);
+        annotation.shape.frames = keyframes;
+        await astore.put(annotation, annotation_id);
+        resolve();
+      };
+      transaction.onerror = (e) => reject(e);
+    });
+  }
+
+  /**
+   * Delete a keyframe from an annotation
+   */
+  async deleteKeyFrame(annotation_id: string, frame: number): Promise<void> {
+    const transaction = this.db.transaction("keyframes", "readwrite");
+    const store = transaction.objectStore("keyframes");
+    const request = store.delete([annotation_id, frame]);
+
+    return new Promise<void>((resolve, reject) => {
+      request.onsuccess = () => resolve();
+      request.onerror = (r) => reject(r);
+    });
+  }
+
+  /**
+   * Get all keyframes for an annotation
+   */
+  async getKeyFrames(annotation_id: string): Promise<VideoFrameSelection[]> {
+    return new Promise<VideoFrameSelection[]>((resolve, reject) => {
+      const transaction = this.db.transaction("keyframes", "readonly");
+      const index = transaction.objectStore("keyframes").index("annotation");
+      const request = index.getAll(IDBKeyRange.only(annotation_id));
+
+      request.onsuccess = () => resolve(request.result);
+      request.onerror = () => reject(request.error);
+    });
+  }
+
+  /**
+   * Get all annotations in a group
+   */
+  async getGroupAnnotations(groupId: string): Promise<VideoAnnotationObject[]> {
+    const transaction = this.db.transaction("annotations", "readonly");
+    const store = transaction.objectStore("annotations");
+
+    const idPromise = new Promise<VideoAnnotationObject | undefined>((resolve, reject) => {
+      const request = store.get(groupId);
+      request.onsuccess = () => resolve(request.result);
+      request.onerror = () => reject(request.error);
+    });
+
+    const groupPromise = new Promise<VideoAnnotationObject[]>((resolve, reject) => {
+      const request = store.index("groupIdIndex").getAll(IDBKeyRange.only(groupId));
+      request.onsuccess = () => resolve(request.result || []);
+      request.onerror = () => reject(request.error);
+    });
+
+    return Promise.all([idPromise, groupPromise]).then(([idResult, groupResults]) => {
+      const allResults = [...groupResults];
+      if (idResult && !allResults.some((r) => r.metadata.id === idResult.metadata.id)) {
+        allResults.push(idResult);
+      }
+      return allResults.sort((a, b) => a.shape.start - b.shape.start);
+    });
+  }
+
+  /**
+   * Update all annotations with a transformation function
+   */
+  async updateAll(updateFn: (annotation: VideoAnnotationObject) => VideoAnnotationObject): Promise<void> {
+    const transaction = this.db.transaction(["annotations"], "readwrite");
+    const objectStore = transaction.objectStore("annotations");
+
+    return new Promise<void>((resolve, reject) => {
+      objectStore.openCursor().onsuccess = (event) => {
+        const cursor = (event.target as IDBRequest)?.result;
+        if (cursor) {
+          try {
+            const updated = updateFn(cursor.value);
+            cursor.update(updated);
+          } catch (error) {
+            reject(error);
+            return;
+          }
+          cursor.continue();
+        }
+      };
+
+      transaction.oncomplete = () => resolve();
+      transaction.onerror = (event) => reject((event.target as IDBRequest)?.error);
+      transaction.onabort = (event) => reject((event.target as IDBRequest)?.error);
+    });
+  }
+}
+
+/**
+ * Initialize IndexedDB and return database instance
+ */
+export async function openAnnotationsDB(name: string, stores: StoresDefinition = storeDefinition) {
+  return new Promise<AnnotationsIndexedDB>((resolve, reject) => {
+    const DBOpenRequest = indexedDB.open(name, currentDBVersion);
+
+    DBOpenRequest.onerror = reject;
+    DBOpenRequest.onsuccess = (_) => {
+      console.info("Database initialized.");
+      const db = new AnnotationsIndexedDB(DBOpenRequest.result);
+      resolve(db);
+    };
+
+    DBOpenRequest.onupgradeneeded = () => {
+      console.info("Upgrading database...");
+      const db = DBOpenRequest.result;
+
+      // Delete all existing object stores
+      Array.from(db.objectStoreNames).forEach((store) => {
+        db.deleteObjectStore(store);
+      });
+
+      // Create object stores and indexes
+      Object.entries(stores).forEach(([store, indexes]) => {
+        const s = db.createObjectStore(store);
+        indexes.forEach((i) => s.createIndex(i.index, i.path, i?.opts));
+      });
+
+      console.info("Database upgraded.");
+    };
+  });
+}
